Safari-611.1.10/LayoutTests/accessibility/mac/progress-with-label-element-expected.txt

label  label wrapping output
This tests that for progress elements, label should be used as accessible name and title should be used as accessible description.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS prog1.title is 'AXTitle: label'
PASS prog1.helpText is 'AXHelp: title'
PASS prog2.title is 'AXTitle: label wrapping output'
PASS successfullyParsed is true

TEST COMPLETE
